+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - eror by opening default repository #138
 - open files with system programm #139
 - use electron.protocol.interceptFileProtocol in place of base64 for images - it's very slow for big files #99
+- save description on blur doesn't work #135
 
 
 ## 0.0.13-alpha